Seperate the overworld into a seperate view
